feat: refactor chapter 4
