+++ algebra/Cargo.toml
@@ -1,8 +1,7 @@
 [package]
-name = "algorithm"
+name = "algebra"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-anyhow = "1.0.96"
-sha2 = "0.10.8"
+sha2 = { workspace = true }
